Move Google login into own file
